Migrate to `buffer_fixed!`
