+++ belt-mac/benches/mod.rs
@@ -1,12 +1,11 @@
 #![feature(test)]
 extern crate test;
 
-use belt_block::BeltBlock;
 use belt_mac::{BeltMac, KeyInit};
 use test::Bencher;
 
 digest::bench_update!(
-    BeltMac::<BeltBlock>::new(&Default::default());
+    BeltMac::new(&Default::default());
     belt_mac_10 10;
     belt_mac_100 100;
     belt_mac_1000 1000;
