+++ cbc-mac/Cargo.toml
@@ -23,9 +23,7 @@ aes = "0.9.0-pre.3"
 des = "0.9.0-pre.3"
 
 [features]
-std = ["digest/std"]
-zeroize = ["cipher/zeroize"]
+zeroize = ["cipher/zeroize", "digest/zeroize"]
 
 [package.metadata.docs.rs]
 all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
