+++ cmac/Cargo.toml
@@ -28,9 +28,7 @@ kuznyechik = "0.9.0-pre.3"
 magma = "0.10.0-pre.3"
 
 [features]
-std = ["digest/std"]
-zeroize = ["cipher/zeroize"]
+zeroize = ["cipher/zeroize", "digest/zeroize"]
 
 [package.metadata.docs.rs]
 all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
