+++ hmac/Cargo.toml
@@ -24,8 +24,7 @@ streebog = { version = "=0.11.0-pre.5", default-features = false }
 hex-literal = "1"
 
 [features]
-std = ["digest/std"]
-reset = [] # Enable ability to reset HMAC instances
+zeroize = ["digest/zeroize"]
 
 [package.metadata.docs.rs]
 all-features = true
