+++ pmac/Cargo.toml
@@ -22,9 +22,7 @@ aes = "0.9.0-pre.3"
 digest = { version = "=0.11.0-pre.10", features = ["dev"] }
 
 [features]
-std = ["digest/std"]
-zeroize = ["cipher/zeroize"]
+zeroize = ["cipher/zeroize", "digest/zeroize"]
 
 [package.metadata.docs.rs]
 all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
